# core/docs/swagger.yaml
basePath: /
definitions:
  api.ReportResponse:
    properties:
      message:
        type: string
      report_id:
        type: string
      timestamp:
        type: string
      type:
        type: string
    type: object
...
  db.CPUStats:
    properties:
      cores:
        type: integer
      load_1:
        type: number
      load_5:
        type: number
      load_15:
        type: number
      usage_percent:
        type: number
    type: object
  db.DiskStats:
    properties:
      free_bytes:
        type: integer
      total_bytes:
        type: integer
      used_bytes:
        type: integer
      used_percent:
        type: number
    type: object
  db.GeoLocation:
    properties:
      city:
        type: string
      country:
        type: string
      hostname:
        type: string
      ip:
        type: string
      loc:
        type: string
      org:
        type: string
      postal:
        type: string
      region:
        type: string
      timezone:
        type: string
    type: object
  db.MemoryStats:
    properties:
      available_bytes:
        type: integer
      free_bytes:
        type: integer
      total_bytes:
        type: integer
      used_bytes:
        type: integer
      used_percent:
        type: number
    type: object
...
  service.AgentReportPayload:
    properties:
      cpu:
        $ref: '#/definitions/db.CPUStats'
      disk:
        $ref: '#/definitions/db.DiskStats'
      location:
        $ref: '#/definitions/db.GeoLocation'
      memory:
        $ref: '#/definitions/db.MemoryStats'
      timestamp:
        type: string
      uptime_seconds:
        type: integer
    type: object
  service.MonitorReportPayload:
    properties:
      error: {}
      health:
        description: up | down
        type: string
      metrics: {}
      timestamp:
        type: string
    required:
    - health
    - metrics
    - timestamp
    type: object
...
  utils.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8999
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Orion Core API for agent management and monitoring
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Orion Core API
  version: "1.0"
paths:
...
  /v1/agents/{agent_id}/{monitor_id}/report:
    post:
      consumes:
      - application/json
      description: Receive and store a monitor report for a specific monitor
      parameters:
      - description: Agent ID
        in: path
        name: agent_id
        required: true
        type: string
      - description: Monitor ID
        in: path
        name: monitor_id
        required: true
        type: string
      - description: Monitor report data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/service.MonitorReportPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Receive monitor report
      tags:
      - reports
...
  /v1/agents/{agent_id}/report:
    post:
      consumes:
      - application/json
      description: Receive and store a system metrics report from an agent
      parameters:
      - description: Agent ID
        in: path
        name: agent_id
        required: true
        type: string
      - description: Agent report data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/service.AgentReportPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/api.ReportResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Receive agent report
      tags:
      - reports
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"